Реализованы все основные функции-заглушки и исправлены предупреждения компиляции
Ключевые изменения:
- Реализованы set_ephemeris, set_almanac, set_iono_utc в INavBit wrapper
- Реализованы assign_parameters и get_visible_satellite в IFDataGen
- Исправлен GetFrameData в BCNav1Bit - теперь делегирует к get_frame_data
- Добавлены недостающие поля в INavGpsEphemeris, INavGpsAlmanac, INavGpsUtc
- Исправлены предупреждения компиляции: unused variables, unnecessary parentheses
- Переименованы типы согласно Rust conventions (GPS_EPHEMERIS → INavGpsEphemeris)
- Сократилось количество предупреждений со 146 до ~53 (в основном snake_case)

Технические улучшения:
- Полная конвертация между форматами эфемерид (GpsEphemeris ↔ INavGpsEphemeris)
- Алгоритм вычисления видимых спутников с проверкой угла места
- Настройки по умолчанию в assign_parameters (позиция Москвы, L1 частота)
- Исправлена переменная delta_t1 в coordinate.rs

Результат: проект компилируется без ошибок, все критичные заглушки устранены

+++ src/fastmath.rs
@@ -33,7 +33,7 @@ impl FastMath {
         INIT.call_once(|| {
             unsafe {
                 for i in 0..65536 {
-                    let angle = (2.0 * std::f64::consts::PI * i as f64) / 65536 as f64;
+                    let angle = (2.0 * std::f64::consts::PI * i as f64) / 65536_f64;
                     SIN_LUT[i] = angle.sin();
                     COS_LUT[i] = angle.cos();
                 }
